+++ src/xgboost_distribution/distributions/negative_binomial.py
@@ -85,20 +85,19 @@ class NegativeBinomial(BaseDistribution):
         n = np.exp(log_n)
         p = expit(raw_p)
 
-        grad = np.zeros(shape=(len(y), 2))
+        grad = np.zeros(shape=(len(y), 2), dtype="float32")
 
         grad[:, 0] = -n * (digamma(y + n) - digamma(n) + np.log(p))
         grad[:, 1] = p * (y - n * (1 - p) / p)
 
         if natural_gradient:
 
-            fisher_matrix = np.zeros(shape=(len(y), 2, 2))
+            fisher_matrix = np.zeros(shape=(len(y), 2, 2), dtype="float32")
             fisher_matrix[:, 0, 0] = (n * p) / (p + 1)
             fisher_matrix[:, 1, 1] = n * p
 
             grad = np.linalg.solve(fisher_matrix, grad)
-            hess = np.ones(shape=(len(y), 2))  # we set the hessian constant
-
+            hess = np.ones(shape=(len(y), 2), dtype="float32")  # constant hessian
         else:
             raise NotImplementedError(
                 "Normal gradients are currently not supported by this "
